ASCII Character Recognition
The purpose of this program is to ask the user to input a series of
characters.  The program will then determine whether or not the
character is a part of the ASCII character set.  If it is not, the
program will disregard the character.  Approved characters will be
printed out in a line until there are 72 characters on a line, at which
point the characters will be printed on the next line.